ui: kill separate agency/admin view components
they were largely similar, this reduces duplication. the `Home`
component now just renders the home page
